refactor(compat): move zod helpers to zodCompat.ts; throw on invalid normalizeRawShapeSchema input; preserve optional in InferRawShape
- Move isZodSchema/isZodRawShape/normalizeRawShapeSchema from standardSchema.ts to a new zodCompat.ts so standardSchema.ts is Standard-Schema-spec only.
- normalizeRawShapeSchema now throws TypeError for inputs that are neither a raw shape nor a Standard Schema, instead of silently returning them.
- InferRawShape now uses z.infer<z.ZodObject<S>> so .optional() fields produce ?: keys.
- Changeset mentions the completable() constraint widening.
